pyproject.toml
[project]
name = "faderport-mcu"
version = "0.1.0"
description = "MIDI bridge translating legacy PreSonus FaderPort to Mackie Control Universal for Logic Pro, Cubase, etc."
requires-python = ">=3.10"
dependencies = ["python-rtmidi>=1.5.0"]

[project.scripts]
faderport-mcu = "faderport_mcu.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
